fix: 중복된 StudyResponseDto 삭제
