+++ src/main/java/com/web/stard/domain/member/api/MemberController.java
@@ -3,11 +3,11 @@ package com.web.stard.domain.member.api;
 import com.web.stard.domain.board.global.application.PostService;
 import com.web.stard.domain.board.global.application.StarScrapService;
 import com.web.stard.domain.board.global.dto.response.PostResponseDto;
-import com.web.stard.domain.board.study.dto.response.StudyResponseDto;
 import com.web.stard.domain.member.application.MemberService;
 import com.web.stard.domain.member.domain.Member;
 import com.web.stard.domain.member.dto.request.MemberRequestDto;
 import com.web.stard.domain.member.dto.response.MemberResponseDto;
+import com.web.stard.domain.study.domain.dto.StudyResponseDto;
 import com.web.stard.global.domain.CurrentMember;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
